Fixes in IsGrowable to get library unpacking back to work; various cosmetic changes; added gamelib to content proj (as backup lib); code to enable loading of backup lib (from file) and backup config (from memory); gamelib updates

+++ IndiegameGarden/IndiegameGarden/Base/GameLibrary.cs
@@ -120,7 +120,7 @@ namespace IndiegameGarden.Base
             GardenItem iggItem = gamesCollection.FindGameNamed("igg");
             if (iggItem != null)
             {
-                iggItem.VisibilityLabel = (GardenConfig.Instance.NewestClientVersion > GardenConfig.Instance.ClientVersion) ? 1 : 0 ;
+                iggItem.VisibilityLabel = (GardenConfig.Instance.NewestClientVersion > GardenConfig.Instance.ClientVersion) ? (byte) 1 : (byte) 0 ;
                 iggItem.Version = GardenConfig.Instance.NewestClientVersion;
             }
         }
